requirements.txt
# =============================================================================
# Movie Recommendation System — Dependencies
# =============================================================================
# Versions are pinned with `~=` (compatible release) for reproducibility:
#   ~=2.2.0 -> >=2.2.0, <2.3.0  (allows patch updates, blocks minor breaking)
# =============================================================================

# ---------- Core data stack ----------
pandas~=2.2.0
numpy~=1.26.0
scipy~=1.12.0

# ---------- Machine learning ----------
scikit-learn~=1.4.0
scikit-surprise~=1.1.3       # SVD / KNN collaborative filtering

# ---------- Visualization ----------
matplotlib~=3.8.0
seaborn~=0.13.0
plotly~=5.20.0               # Interactive charts + choropleth maps

# ---------- Web / API ----------
requests~=2.31.0
python-dotenv~=1.0.0         # .env file support
tqdm~=4.66.0                 # Progress bars (TMDB fetch, training)

# ---------- Dashboard ----------
streamlit~=1.32.0

# ---------- Notebook environment ----------
jupyter~=1.0.0
ipykernel~=6.29.0
nbformat~=5.10.0

# ---------- Performance helpers ----------
pyarrow~=15.0.0              # Fast parquet/CSV engine for pandas

# =============================================================================
# NOTE on geopandas:
# The original spec mentioned `geopandas`, but for choropleth maps we use
# plotly's built-in country support (ISO-3166 alpha-3 codes), which avoids
# heavy GDAL/Fiona/PROJ system dependencies and works on Streamlit Cloud.
# Add `geopandas~=0.14.0` here only if shapefile-level work is needed.
# =============================================================================
